{"schema":"decodex/commit/1","summary":"Split core block validation modules","authority":"manual"}
